requirements.txt
Daphne
Django
djangorestframework
djangorestframework-simplejwt
channels
django-cors-headers
channels_redis
